Rename create_counter to counter and reduce number time it calls on headers
